pyproject.toml
[project]
name = "temporal-ecommerce-agent"
version = "0.1.0"
description = "Multi-Agent E-commerce Order Processing with Temporal and OpenAI Agents SDK"
authors = [
    {name = "Demo Developer", email = "demo@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "temporalio>=1.5.0",
    "openai-agents>=0.1.0",
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 